Sped up the Cache slightly
I realized that the library retrieves cached items much more often
than it creates them so now the exact expiry time is calculated
when an item is stored and all the getter has to do is check "X < Y".

Looped the `list_monitors_info` function 10 million times and got
a 2 second speed boost so that's progress at least.