Re-wrote top-level get/set_brightness functions
The idea for this update will be to speed things up by making lower-level functions less convenient, requiring less decisions.
So all low-level brightness set/getters now require the display kwarg as int so that they don't have to call filter_monitors, something
which the top-level functions can sort out.

Because of this, many low-level convenience functions are deprecated. This includes:
* [linux/windows].set_brightness
* [linux/windows].get_brightness
* [linux/windows].list_monitors
* windows.[WMI/VCP].get_display_names
* linux.[Light/XRandr/DDCUtil].get_display_names

More changes will be made to low-level functions in the future in the hopes of speeding things up