Fix AttributeError in list_monitors_info
Changed order in which functions/classes are defined. Now the most relevant functions are at the top of the file
(eg: get_brightness, set_brightness...) and the private/irrelevant ones are at the bottom (eg: flatten_list, __Cache...).
This is because the new documentation system sorts functions by order of appearance when generating the documentation