Change test structure to make debugging error messages easier.
Previously, a single class would test both methods. If something went wrong the error would say something like
"error in class 'TestAllMethodsBasic'..." and it was hard to find out which method, `WMI` or `VCP`, the error was
coming from.

Now there is a generic `BasicMethodTest` class that provides the generic tests that `TestWMI` and `TestVCP` inherit.