Add toolchain for building documentation.
I intend to use this to deploy to GitHub pages in the future once a few more wrinkles are worked out.
In the mean time, this can be used to build local copies of the documentation.

The "--version-navigator" switch for `tools/make_docs.py` is to enable the version navigation javascript
that I will be using for the GitHub pages deployment in the future but isn't all that useful to the average user