Add support for building documentation for other versions of module.
Specify the path to the source code of the version (via the '--path' flag) you wish to generate documentation for and
the script should automatically discover the version of that code and generate the documentation.

Pass the '--clean' flag to remove all existing documentation from the directory.

Example to generate docs for v0.5.1 whilst keeping any existing documentation:
python docs/make.py --path archives/v0.5.1/screen_brightness_control

Note that you must store the code in a path that ends with the folder "screen_brightness_control" so that pdoc picks up
the correct module name