Only show highest patch level by default.
Previously, the version navigator would take up alot of space showing ALL of the patch levels
for each release. Now it will show the latest patch level for each release with earlier levels
inside a dropdown menu, which expands upon clicking the "+" button.

This is an improvement over the previous implementation used where the dropdown would expand upon mouseover, making
it harder to browse to the specific version you want as the dropdowns displace the other versions up/down the page as you mouse over them.