Deprecate `linux.XBacklight` brightness method for removal after next release.
The main reason for this is that, as far as I can tell, the xbacklight executable does not support addressing individual
displays and is the only brightness method with this problem.

That makes it a pain to support as it cannot be used with the `display` kwarg and can only be used as the `method` kwarg for
some functions, creating alot of situations where I need to build in a few extra if/else clauses to deal with these edge cases.

Hopefully with the recent introductions of the `SysFiles` and `I2C` classes, this shouldn't cause any issues for display compatibility.