Improve `SysFiles` and `Light` methods display compatibility.
Previously, only displays that had an available `/sys/class/backlight/.../device/edid` file were allowed,
as that was the only meaningful way of differentiating the displays. It also filtered out the `acpi_video0`
displays when `intel_backlight` was available.

This approach doesn't work on displays that do not have this EDID file available, for example, my Pinebook
Pro on Manjaro.

The new approach now is to group and differentiate displays based on the subsystems in
`/sys/class/backlight/.../subsystem`. Once grouped, the display with the highest resolution
scale is then used to gather all of the other available information.