Implement logarithmic brightness scaling in `fade_brightness`.
As `fade_brightness` iterates from (for example) 0 to 100 it used to set the brightness for every step in between. Thats 101 calls to `set_brightness`.
Using a logarithmic scale it can skip most of the higher percentages and achieve the same fade in 54 calls.

This is cool because single-percent brightness changes are not as noticeable at the higher end.

The scaling follows the curve `y=10^(x/50)` (which is the inverse of `y=50log(x)` hence why I call it logarithmic).