Add "fast mode" testing to `test_init.py`.
In the test helpers file there is a new `FakeMethodTest` class which emulates the behaviour of a regular brightness method,
providing a `get_display_info` and brightness set/get functions. The `get_display_info` will get and then cache real display
info from the test machine. The brightness set/get functions then just keep track of the brightness in a class dictionary.

Then, when tests are set up, the `sbc.get_methods` function is overridden to only return the phony `FakeMethodTest` brightness
method, tricking the library into ignoring the real (slow) OS specific methods.

The point of this feature is to speed up testing for cases where we are testing the function's logic and NOT the function's
purpose.
For example, when testing `filter_monitors` I only care that it can filter a list of monitors correctly, so why waste time making
dozens of calls to `list_monitors_info`? Why not just give it a cached list to filter?

You can test using fast mode by passing the `--test-fast` flag to `test_init.py`.
I recommend using this for small changes and bug fixes but using the regular, slow testing for major releases.