Add function to return info useful for debugging.
Often, when an issue is raised on this repo, I'll end up asking for X info
and Y command output and Z code result. I imagine this is quite frustrating for
whoever just wants their bug fixed.

This new function should allow me to ask for 1 thing and have almost all of the
information I need to start fixing issues.