Remove platform check for `fcntl` import.
This platform check was required in the past to allow `pdoc` to import `screen_brightness_control/linux.py`
in order to generate documentation for it. However, 3c736e2 enabled building documentation regardless
of the platform, so this check should no longer be necessary.

Also I usually build most of the docs on linux anyway