Fix `logarithmic_range` skipping lower numbers when iterating backwards.
The idea of `logarithmic_range` is, as explained in 7b3704b, to skip many of the higher brightness percentages
as 99% -> 100% is a 0.01 increase in brightness whereas 1%-2% is double the brightness.

There was a logic bug where, when iterating backwards, the function would skip many of the lower values close
to 0, instead of the higher values approaching 100. This commit fixes that bug by mirroring the logarithmic
curve when we are iterating backwards