Remove redundant type info from docstrings (1/4)
All of this library's functions are type hinted, so there is little point recording
types in the docstrings. The new approach will be to document types via hints and
explain the parameters in the docstrings.

Also added a `Percentage` type alias (name may change) to refer to the int/str
absolute/relative brightness levels many functions accept