Make TTL cache delete expired data automatically.
Upgraded the `expire` method to also remove all data that has expired from the cache.
Previously, if data wasn't explicity deleted, it could stagnate and stay there forever. Now the `expire` method is called during the
`get` method, so this should no longer be an issue.

Also tweaked `set` so that data is no longer stored if the cache is disabled