Upgrade documentation out of date notice to redirect to correct page on latest docs.
Previous behaviour was that it would just redirect to the root of the latest documentation, which meant you
lost your place if you were looking at a specific item on a submodule.

Now it will attempt to piece together a new URL using the latest version and whatever URL fragment you're looking at.
It will check if that URL returns 200 and redirect there if so. Otherwise will fall back to the old behaviour.

This is a bit slower and more expensive than the old way but it only happens on old docs so shouldn't be an issue.

Also removed docs/ dir from testing CICD because why was that there in the first place