Replace old `unittest` test suite with new `pytest` test suite.
This has been a long time in the making but the test suite has been (*mostly*) ported over to pytest.
The test for `windows.py` has not been yet been ported. This isn't a huge loss as the old one wasn't able to run in CI/CD anyway and never got used.

The rationale for the port is somewhat covered in 89642fc. Test coverage is now much better, testing is more in depth and we are finally
able to test lower level code (eg: `linux.py`) in a reasonable manner (mocking displays rather than hammering real displays). This means
that tests are fast, reliable and much more fun to write!

I'm hoping to add tests for `windows.py` in the near future which will make the project near 100% coverage for the first time ever.