Fix excessive calls to `dxva2.DestroyPhysicalMonitor`
`iter_physical_monitors` handles cleanup automatically, unless you break out of the generator.
This meant that when `display == None`, breaks weren't happening and we were having extra calls
to `DestroyPhysicalMonitor`.

Furthermore, `VCP.set_brightness` was not breaking out of the generator when the desired index
had been reached, resulting in unnecessary iterations of the generator