Add UIDs to all Linux brightness methods
The UID here is based on the I2C path of the display, which is easily exposed for I2C and DDCutil.
For SysFiles and XRandr, the process is a bit more complicated. It relies on the idea that we can
link the entries in `/sys/class/backlight` to `/sys/class/drm` which should contain a link to the
I2C system.

For SysFiles, we expand the `/sys/class/backlight/*/device` symlink to expose the path to the
underlying display connection (in this case, usually eDP). We can follow this link to the
equivalent path in `/sys/class/drm` which will usually have an `i2c-X` or `ddc/i2c-dev` folder
inside that maps to a relevant I2C connection.

On XRandr, it's a similar process, except we start by looking in `/sys/class/drm` and matching the
connection names to our interface name. For example: HDMI-1 -> card0-HDMI-A-1 or DP-1 -> card0-DP-1.
We can then find the I2C connection by browsing the folder tree.

I have tested this on multiple devices and in all cases, both laptop and desktop displays were able
to be uniquely identified in this manner across brightness methods. This means that displays can
avoid being duplicated across all methods even if there is no EDID information available.

There are a few cases in which this doesn't work. On my Pinebook Pro, the edp-backlight subsystem
seems to be a completely different subsystem to the actual display, with no tangible connection
between the two. In this case, a UID cannot be found.
On a laptop running a Fedora Wayland live CD, there was no `intel_backlight` folder, which meant
that no UID was found. On the same laptop running X11 mint, `intel_backlight` was present and the
UID was found so I believe this is a driver/kernel module issue.