Add `COM_FLAG` constant for controlling how COM is initialised.
Issue #42 shows issues with initialising WMI in a pynput app. This appears to be down to how COM is initialised.
Using MTA in this case appears to solve the error, but pythoncom defaults to STA and I don't know the ramifications
of defaulting to MTA.

I'm leaving the default behavious as is and adding a flag that users can change if they wish