Deprecate `XRandr` method
XRandr will be removed in the next release.
There are a couple of reasons to deprecate XRandr:

1. X11 only
Xrandr doesn't support Wayland, and X11 is EOL. Gnome, Fedora and Ubuntu have dropped X11, and KDE is going to sometime soon.
70% of KDE users are on Wayland anyway.

2. It doesn't change backlight, only gamma
This has been a longstanding issue which is that xrandr doesn't change the backlight, it applies a gamma filter. This meant you could
accidentally dim your screen "twice" with different brightness methods and not realise how to fix it. This is confusing behaviour
and isn't what this library should do.

3. Functionality has been superseded
XRandr simply isn't needed any more. The SysFiles and I2C classes both cover pretty much all displays, with DDCUtil covering USB displays.
It's just not required anymore.

I'll probably publish a release with the deprecation warning in place and then remove it in the next one, or in a few months, whichever comes first